+++ core/r4300/compare_core.cpp
@@ -36,30 +36,30 @@
 #include <shared/services/LoggingService.h>
 
 static FILE* f;
-static int pipe_opened = 0;
-static long long int comp_reg[32];
-extern unsigned long interp_addr;
-static unsigned long old_op;
-int compare_core_mode = 0;
+static int32_t pipe_opened = 0;
+static int64_t comp_reg[32];
+extern uint32_t interp_addr;
+static uint32_t old_op;
+int32_t compare_core_mode = 0;
 
 void display_error(const char* txt)
 {
-    int i;
-    unsigned long* comp_reg2 = (unsigned long*)comp_reg;
+    int32_t i;
+    uint32_t* comp_reg2 = (uint32_t*)comp_reg;
     g_core_logger->error("err: {}", txt);
     if (interpcore)
     {
-        g_core_logger->info("addr:{:#06x}", (int)interp_addr);
-        if (!strcmp(txt, "PC")) g_core_logger->info("{:#06x} - {:#06x}", (int)interp_addr, *(int*)&comp_reg[0]);
+        g_core_logger->info("addr:{:#06x}", (int32_t)interp_addr);
+        if (!strcmp(txt, "PC")) g_core_logger->info("{:#06x} - {:#06x}", (int32_t)interp_addr, *(int32_t*)&comp_reg[0]);
     }
     else
     {
-        g_core_logger->info("addr:{:#06x}", (int)PC->addr);
-        if (!strcmp(txt, "PC")) g_core_logger->info("{:#06x} - {:#06x}", (int)PC->addr, *(int*)&comp_reg[0]);
+        g_core_logger->info("addr:{:#06x}", (int32_t)PC->addr);
+        if (!strcmp(txt, "PC")) g_core_logger->info("{:#06x} - {:#06x}", (int32_t)PC->addr, *(int32_t*)&comp_reg[0]);
     }
-    g_core_logger->info("{:#06x}, {:#06x}", (unsigned int)reg_cop0[9], (unsigned int)comp_reg2[9]);
-    g_core_logger->info("erreur @:{:#06x}", (int)old_op);
-    g_core_logger->info("erreur @:{:#06x}", (int)vr_op);
+    g_core_logger->info("{:#06x}, {:#06x}", (uint32_t)reg_cop0[9], (uint32_t)comp_reg2[9]);
+    g_core_logger->info("erreur @:{:#06x}", (int32_t)old_op);
+    g_core_logger->info("erreur @:{:#06x}", (int32_t)vr_op);
 
     if (!strcmp(txt, "gpr"))
     {
@@ -76,7 +76,7 @@ void display_error(const char* txt)
         {
             if (reg_cop0[i] != comp_reg2[i])
                 g_core_logger->info("reg_cop0[{}]={:#06x} != reg_cop0[{}]={:#06x}",
-                       i, (unsigned int)reg_cop0[i], i, (unsigned int)comp_reg2[i]);
+                       i, (uint32_t)reg_cop0[i], i, (uint32_t)comp_reg2[i]);
         }
     }
     /*for (i=0; i<32; i++)
@@ -106,7 +106,7 @@ void check_input_sync(unsigned char* value)
 
 void compare_core()
 {
-    //static int wait=1;
+    //static int32_t wait=1;
 
     if (compare_core_mode == 0)
     {
@@ -147,7 +147,7 @@ void compare_core()
         /*if(wait == 1 && reg_cop0[9] > 0x35000000) wait=0;
         if(wait) return;*/
 
-        fread(comp_reg, 1, sizeof(long), f);
+        fread(comp_reg, 1, sizeof(int32_t), f);
         if (feof(f))
         {
             pipe_opened = 0;
@@ -167,18 +167,18 @@ void compare_core()
                 display_error("PC");
         }
         /*
-        fread (comp_reg, 32, sizeof(long long int), f);
-        if (memcmp(reg, comp_reg, 32*sizeof(long long int)))
+        fread (comp_reg, 32, sizeof(int64_t), f);
+        if (memcmp(reg, comp_reg, 32*sizeof(int64_t)))
           display_error("gpr");
-        fread (comp_reg, 32, sizeof(long), f);
-        if (memcmp(reg_cop0, comp_reg, 32*sizeof(long)))
+        fread (comp_reg, 32, sizeof(int32_t), f);
+        if (memcmp(reg_cop0, comp_reg, 32*sizeof(int32_t)))
           display_error("cop0");
-        fread (comp_reg, 32, sizeof(long long int), f);
-        if (memcmp(reg_cop1_fgr_64, comp_reg, 32*sizeof(long long int)))
+        fread (comp_reg, 32, sizeof(int64_t), f);
+        if (memcmp(reg_cop1_fgr_64, comp_reg, 32*sizeof(int64_t)))
           display_error("cop1");
           */
-        /*fread(comp_reg, 1, sizeof(long), f);
-        if (memcmp(&rdram[0x31280/4], comp_reg, sizeof(long)))
+        /*fread(comp_reg, 1, sizeof(int32_t), f);
+        if (memcmp(&rdram[0x31280/4], comp_reg, sizeof(int32_t)))
           display_error("mem");*/
         /*fread (comp_reg, 4, 1, f);
         if (memcmp(&FCR31, comp_reg, 4))
@@ -187,7 +187,7 @@ void compare_core()
     }
     else
     {
-        //if (reg_cop0[9] > 0x6800000) g_core_logger->info("PC={:#06x}", (int)PC->addr);
+        //if (reg_cop0[9] > 0x6800000) g_core_logger->info("PC={:#06x}", (int32_t)PC->addr);
         if (pipe_opened != 2)
         {
             if (f) fclose(f);
@@ -198,15 +198,15 @@ void compare_core()
         if(wait) return;*/
 
         if (interpcore)
-            fwrite(&interp_addr, 1, sizeof(long), f);
+            fwrite(&interp_addr, 1, sizeof(int32_t), f);
         else
-            fwrite(&PC->addr, 1, sizeof(long), f);
+            fwrite(&PC->addr, 1, sizeof(int32_t), f);
         /*
-        fwrite(reg, 32, sizeof(long long int), f);
-        fwrite(reg_cop0, 32, sizeof(long), f);
-        fwrite(reg_cop1_fgr_64, 32, sizeof(long long int), f);
+        fwrite(reg, 32, sizeof(int64_t), f);
+        fwrite(reg_cop0, 32, sizeof(int32_t), f);
+        fwrite(reg_cop1_fgr_64, 32, sizeof(int64_t), f);
         */
-        //fwrite(&rdram[0x31280/4], 1, sizeof(long), f);
+        //fwrite(&rdram[0x31280/4], 1, sizeof(int32_t), f);
         /*fwrite(&FCR31, 4, 1, f);*/
     }
 }
