+++ core/r4300/recomp.h
@@ -42,21 +42,21 @@ typedef struct _precomp_instr
     {
         struct
         {
-            long long int* rs;
-            long long int* rt;
-            short immediate;
+            int64_t* rs;
+            int64_t* rt;
+            int16_t immediate;
         } i;
 
         struct
         {
-            unsigned long inst_index;
+            uint32_t inst_index;
         } j;
 
         struct
         {
-            long long int* rs;
-            long long int* rt;
-            long long int* rd;
+            int64_t* rs;
+            int64_t* rt;
+            int64_t* rd;
             unsigned char sa;
             unsigned char nrd;
         } r;
@@ -65,7 +65,7 @@ typedef struct _precomp_instr
         {
             unsigned char base;
             unsigned char ft;
-            short offset;
+            int16_t offset;
         } lf;
 
         struct
@@ -79,31 +79,31 @@ typedef struct _precomp_instr
 #endif
     } f;
 
-    unsigned long addr;
-    unsigned long local_addr;
+    uint32_t addr;
+    uint32_t local_addr;
     reg_cache_struct reg_cache_infos;
     void (*s_ops)();
-    unsigned long src;
+    uint32_t src;
 } precomp_instr;
 
 typedef struct _precomp_block
 {
     precomp_instr* block;
-    unsigned int start;
-    unsigned int end;
+    uint32_t start;
+    uint32_t end;
     unsigned char* code;
-    unsigned int code_length;
-    unsigned int max_code_length;
+    uint32_t code_length;
+    uint32_t max_code_length;
     void* jumps_table;
-    int jumps_number;
+    int32_t jumps_number;
     //unsigned char md5[16];
-    unsigned long adler32;
+    uint32_t adler32;
 } precomp_block;
 
-void recompile_block(long* source, precomp_block* block, unsigned long func);
-void init_block(long* source, precomp_block* block);
+void recompile_block(int32_t* source, precomp_block* block, uint32_t func);
+void init_block(int32_t* source, precomp_block* block);
 void recompile_opcode();
-void prefetch_opcode(unsigned long op);
+void prefetch_opcode(uint32_t op);
 void dyna_jump();
 void dyna_start(void (*code)());
 void dyna_stop();
