+++ core/r4300/vcr.h
@@ -17,7 +17,7 @@ enum
 
 #pragma pack(push, 1)
 /**
- * The movie flag bitfield for extended movies.  
+ * The movie flag bitfield for extended movies.
  */
 typedef union ExtendedMovieFlags
 {
@@ -72,27 +72,27 @@ typedef struct MovieHeader
     /**
      * \brief <c>M64\0x1a</c>
      */
-    unsigned long magic;
+    uint32_t magic;
 
     /**
      * \brief Currently <c>3</c>
      */
-    unsigned long version;
+    uint32_t version;
 
     /**
      * \brief Movie creation time, used to correlate savestates to a movie
      */
-    unsigned long uid;
+    uint32_t uid;
 
     /**
      * \brief Amount of VIs in the movie
      */
-    unsigned long length_vis;
+    uint32_t length_vis;
 
     /**
      * \brief The low word of the amount of loaded states during recording.
      */
-    unsigned long rerecord_count;
+    uint32_t rerecord_count;
 
     /**
      * \brief The VIs per second of the rom this movie was recorded on
@@ -109,7 +109,7 @@ typedef struct MovieHeader
     unsigned char num_controllers;
 
     /**
-     * The extended format version. Old movies have it set to <c>0</c>.  
+     * The extended format version. Old movies have it set to <c>0</c>.
      */
     uint8_t extended_version = 1;
 
@@ -121,20 +121,20 @@ typedef struct MovieHeader
     /**
      * \brief Amount of input samples in the movie, ideally equal to <c>(file_length - 1024) / 4</c>
      */
-    unsigned long length_samples;
+    uint32_t length_samples;
 
     /**
      * \brief What state the movie is expected to start from
      * \remarks vcr.h:32
      */
-    unsigned short startFlags;
+    uint16_t startFlags;
 
-    unsigned short reserved2;
+    uint16_t reserved2;
 
     /**
      * \brief Bitfield of flags for each controller. Sequence of present, mempak and rumble bits with per-controller values respectively (c1,c2,c3,c4,r1,...)
      */
-    unsigned long controller_flags;
+    uint32_t controller_flags;
 
     /**
      * The extended movie data. Only valid if <c>extended_version != 0</c>.
@@ -161,12 +161,12 @@ typedef struct MovieHeader
     /**
      * \brief The primary checksum of the rom the movie was recorded on
      */
-    unsigned long rom_crc1;
+    uint32_t rom_crc1;
 
     /**
      * \brief The rom country of the rom the movie was recorded on
      */
-    unsigned short rom_country;
+    uint16_t rom_country;
 
     char reserved_bytes[56];
 
@@ -225,11 +225,11 @@ enum class e_warp_modify_status
  */
 struct t_movie_freeze
 {
-    unsigned long size;
-    unsigned long uid;
-    unsigned long current_sample;
-    unsigned long current_vi;
-    unsigned long length_samples;
+    uint32_t size;
+    uint32_t uid;
+    uint32_t current_sample;
+    uint32_t current_vi;
+    uint32_t length_samples;
     std::vector<BUTTONS> input_buffer;
 };
 
@@ -241,7 +241,7 @@ bool task_is_recording(e_task task);
  * \param index The polled controller's index
  * \param input The controller's input data
  */
-void vcr_on_controller_poll(int index, BUTTONS* input);
+void vcr_on_controller_poll(int32_t index, BUTTONS* input);
 
 /**
  * \brief Notifies VCR engine about a new VI
@@ -314,7 +314,7 @@ namespace VCR
      *	"n" - Frame
      *	"+n", "-n" - Relative to current sample
      *	"^n" - Sample n from the end
-     *	
+     *
      */
     CoreResult begin_seek(std::string str, bool pause_at_end);
 
@@ -388,17 +388,17 @@ namespace VCR
      * Then, the closest savestate prior to that sample will be loaded and recording will be resumed with the modified inputs up to the sample the function was called at.
      *
      * This operation is long-running and status is reported via the WarpModifyStatusChanged message.
-     * A successful warp modify operation can be detected by the status changing from warping to none with no errors inbetween. 
-     * 
+     * A successful warp modify operation can be detected by the status changing from warping to none with no errors inbetween.
+     *
      * If the provided buffer is identical to the current input buffer (in both content and size), the operation will succeed with no seek.
-     * 
+     *
      * If the provided buffer is larger than the current input buffer and the first differing input is after the current sample, the operation will succeed with no seek.
      * The input buffer will be overwritten with the provided buffer and when the modified frames are reached in the future, they will be "applied" like in playback mode.
      *
      * If the provided buffer is smaller than the current input buffer, the VCR engine will seek to the last frame and otherwise perform the warp modification as normal.
      *
      * An empty input buffer will cause the operation to fail.
-     * 
+     *
      * \param inputs The input buffer to use.
      * \return The operation result
      */
@@ -417,7 +417,7 @@ namespace VCR
 
     /**
      * Gets whether a seek savestate exists at the specified frame.
-     * The returned state changes when the <c>SeekSavestatesChanged</c> message is sent. 
+     * The returned state changes when the <c>SeekSavestatesChanged</c> message is sent.
      */
     bool has_seek_savestate_at_frame(size_t frame);
